backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vladimir Mashyanov <vova.mashyanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.2"
uvicorn = { extras = ["standard"], version = "^0.31.1" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
jinja2 = "^3.1.4"
sqlalchemy = "^2.0.36"
aiosqlite = "^0.20.0"
pydantic-settings = "^2.6.1"
python-jose = "^3.3.0"
bcrypt = "^4.2.1"
mypy = "^1.13.0"
httpx = "^0.28.0"
pytest = "^8.3.3"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
